refactor: change folders and small changes from comments Pull Request
