+++ src/context/index.ts
@@ -1,0 +1,1 @@
+export { LanguageContext } from './LanguageContext';
