+++ src/hooks/useLanguage.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { LanguageContext } from '../context';
+
+export const useLanguage = () => useContext(LanguageContext);
