+++ src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import { ErrorBoundary } from './components/ErrorBoundary';
 import { PathRouter } from './routes';
-import { LanguageProvider } from './utils/context/LanguageProvider';
+import { LanguageProvider } from './providers/LanguageProvider';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
