+++ src/routes/ErrorRoute/ErrorRoute.tsx
@@ -1,13 +1,13 @@
 import { useRouteError } from 'react-router-dom';
 import { isRouteErrorResponse } from 'react-router-dom';
-import { useLanguage } from '../../utils/context';
+import { useLanguage } from '../../hooks';
 
 export function ErrorRoute() {
   const { dictionary } = useLanguage();
   const error: unknown = useRouteError();
   return (
     <>
-      <h1>{dictionary.oops}</h1>
+      <h1>{dictionary.errorRoute}</h1>
       {isRouteErrorResponse(error) && (
         <>
           <h2>{error.status}</h2>
