--- src/utils/context/index.ts
@@ -1,1 +1,0 @@
-export { LanguageContext, useLanguage } from './LanguageContext';
